scripts/requirements.txt
pyjwt
PyGithub
cryptography
click
requests

python-dotenv
